requirements.txt
# Database Connectors
aiomysql>=0.1.1
asyncpg>=0.28.0
motor>=3.3.0
pymongo>=4.5.0

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0

# Web Framework (for future phases)
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0